# scaffolds/blueprints/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/organizations/{id}": {
      "get": {
        "description": "Fetch detailed account information.",
        "operationId": "OrganizationController_fetchOrganization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Organization ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The organization details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchOrganizationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Failed to get organization details"
          },
          "404": {
            "description": "Organization not found"
          }
        },
        "summary": "Fetch organization details",
        "tags": [
          "Organizations"
        ]
      },
      "patch": {
        "description": "Update the organization details.",
        "operationId": "OrganizationController_updateOrganization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Organization ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The organization has been successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchOrganizationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Failed to update organization"
          },
          "401": {
            "description": "User does not have permission to update organizations"
          },
          "404": {
            "description": "Organization not found"
          }
        },
        "summary": "Update organization",
        "tags": [
          "Organizations"
        ]
      }
    },
    "/api/organizations": {
      "post": {
        "description": "Create a new organization for an account.",
        "operationId": "OrganizationController_createOrganization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The organization has been successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchOrganizationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Failed to create organization"
          },
          "401": {
            "description": "User does not have permission to create organizations"
          }
        },
        "summary": "Create organization",
        "tags": [
          "Organizations"
        ]
      }
    },
...
  },
  "info": {
    "title": "SaaSFoundry API",
    "description": "An open-source solution for managing clients, invoices, and financial tasks.",
    "version": "1.0.0",
    "contact": {
      "name": "SaaSFoundry Team",
      "url": "https://github.com/agachet/saasfoundry",
      "email": "anthony.gachet@diamondforge.fr"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FetchOrganizationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Organization unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Organization name",
            "example": "Acme Corporation"
          },
          "type": {
            "type": "string",
            "description": "Organization type",
            "example": "COMPANY",
            "enum": [
              "COMPANY",
              "ASSOCIATION",
              "COMMUNITY"
            ]
          },
          "description": {
            "type": "object",
            "description": "Organization description",
            "example": "A leading technology company",
            "nullable": true
          },
          "website": {
            "type": "object",
            "description": "Organization website",
            "example": "https://www.acme.com",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Organization creation date",
            "example": "2024-03-06T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Organization last update date",
            "example": "2024-03-06T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Organization name",
            "example": "Acme Corporation"
          },
          "type": {
            "type": "string",
            "description": "Organization type",
            "example": "COMPANY",
            "enum": [
              "COMPANY",
              "ASSOCIATION",
              "COMMUNITY"
            ]
          },
          "accountId": {
            "type": "string",
            "description": "Account ID that this organization belongs to",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "description": {
            "type": "string",
            "description": "Organization description",
            "example": "A leading technology company"
          },
          "website": {
            "type": "string",
            "description": "Organization website",
            "example": "https://www.acme.com"
          }
        },
        "required": [
          "name",
          "type",
          "accountId"
        ]
      },
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Organization name",
            "example": "Acme Corporation"
          },
          "type": {
            "type": "string",
            "description": "Organization type",
            "example": "COMPANY",
            "enum": [
              "COMPANY",
              "ASSOCIATION",
              "COMMUNITY"
            ]
          },
          "description": {
            "type": "string",
            "description": "Organization description",
            "example": "A leading technology company"
          },
          "website": {
            "type": "string",
            "description": "Organization website",
            "example": "https://www.acme.com"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
...
    }
  }
}